Criando cliente para baixar arquivos json
