+++ client/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:3-alpine
+
+WORKDIR /
+
+COPY client.py client.py
+COPY requirements.txt requirements.txt
+
+RUN pip install -r requirements.txt
+
+ENTRYPOINT [ "python", "client.py" ]
