+++ nginx/Dockerfile
@@ -1,0 +1,3 @@
+FROM nginx:alpine
+
+RUN chmod -R 777 '/usr/share/nginx/html/'
